Moved common-prop-types out of src/ folder (#7871)
