+++ app/javascript/Snackbar/SnackbarItem.jsx
@@ -1,7 +1,7 @@
 import { h } from 'preact';
 import PropTypes from 'prop-types';
 import { Button } from '@crayons';
-import { defaultChildrenPropTypes } from '../src/components/common-prop-types';
+import { defaultChildrenPropTypes } from '../common-prop-types';
 
 export const snackbarItemProps = {
   children: defaultChildrenPropTypes.isRequired,
