+++ app/javascript/articles/PodcastArticle.jsx
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import { articlePropTypes } from '../src/components/common-prop-types/article-prop-types';
+import { articlePropTypes } from '../common-prop-types/article-prop-types';
 
 export const PodcastArticle = ({ article }) => {
   return (
