+++ app/javascript/articles/components/TagList.jsx
@@ -1,9 +1,9 @@
 import { h } from 'preact';
-import { tagPropTypes } from '../../src/components/common-prop-types';
+import { tagPropTypes } from '../../common-prop-types';
 
 export const TagList = ({ tags = [] }) => (
   <div className="crayons-story__tags">
-    {tags.map(tag => (
+    {tags.map((tag) => (
       <a className="crayons-tag" href={`/t/${tag}`}>
         <span className="crayons-tag__prefix">#</span>
         {tag}
