+++ app/javascript/common-prop-types/article-prop-types.js
