+++ app/javascript/common-prop-types/default-children-prop-types.js
