+++ app/javascript/common-prop-types/selected-tags-prop-types.js
