+++ app/javascript/common-prop-types/user-prop-types.js
