+++ app/javascript/crayons/ButtonGroup/ButtonGroup.jsx
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import { defaultChildrenPropTypes } from '../../src/components/common-prop-types';
+import { defaultChildrenPropTypes } from '../../common-prop-types';
 
 export const ButtonGroup = ({ children }) => (
   <div className="crayons-btn-group">{children}</div>
