+++ app/javascript/crayons/formElements/FormField/FormField.jsx
@@ -1,6 +1,6 @@
 import { h } from 'preact';
 import PropTypes from 'prop-types';
-import { defaultChildrenPropTypes } from '../../../src/components/common-prop-types';
+import { defaultChildrenPropTypes } from '../../../common-prop-types';
 
 // Only radio and checkboxes require an additional CSS class (variant class). Other form elements do not.
