+++ app/javascript/crayons/storybook-utilities/Fieldset.jsx
@@ -1,5 +1,5 @@
 import { h } from 'preact';
-import { defaultChildrenPropTypes } from '../../src/components/common-prop-types';
+import { defaultChildrenPropTypes } from '../../common-prop-types';
 
 export const Fieldset = ({ children }) => (
   <fieldset style={{ border: 'none' }}>{children}</fieldset>
